pyproject.toml
[project]
name = "cv-agro"
version = "1.0.0"
description = "Plant disease detection and leaf area quantification using PlantVillage"
requires-python = ">=3.10"

dependencies = [
    # Core ML
    "torch>=2.0.0",
    "torchvision>=0.15.0",
    "timm>=0.9.0",

    # Dataset
    "datasets>=2.18.0",
    "huggingface-hub>=0.20.0",

    # Computer Vision
    "opencv-python>=4.8.0",
    "Pillow>=10.0.0",

    # Numerics / ML utilities
    "numpy>=1.24.0",
    "scikit-learn>=1.3.0",

    # Visualisation
    "matplotlib>=3.7.0",
    "seaborn>=0.12.0",

    # Progress / utilities
    "tqdm>=4.65.0",
]

[project.optional-dependencies]
notebooks = [
    "jupyter>=1.0.0",
    "ipywidgets>=8.0.0",
]

[project.scripts]
train   = "train:main"
predict = "predict:main"

[build-system]
requires      = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.uv]
# Pin to CPU-only torch to avoid pulling in CUDA wheels on machines without a GPU.
# Remove or change the index to get CUDA builds (see comment below).
#
# For CUDA 12.1:
#   [[tool.uv.index]]
#   name = "pytorch-cu121"
#   url  = "https://download.pytorch.org/whl/cu121"
#   explicit = true
#
# [tool.uv.sources]
#   torch      = [{ index = "pytorch-cu121" }]
#   torchvision = [{ index = "pytorch-cu121" }]

[[tool.uv.index]]
name     = "pytorch-cpu"
url      = "https://download.pytorch.org/whl/cpu"
explicit = true

[tool.uv.sources]
torch       = [{ index = "pytorch-cpu" }]
torchvision = [{ index = "pytorch-cpu" }]
